Moved the landing-page from components to views
